Move pathpolicy out of internal/ into pkg/
Following up from 41b1d00770061f616bc7defc53a03cb711227e90, we now move
the pathpolicy package out of internal as well.  This is mostly
necessary because there are public functions and types, that depend
on types in pathpolicy.

blueprint.CheckMountpointsPolicy() takes *pathpolicy.PathPolicies as a
second argument.
The policies under pkg/policies/ are created using
pathpolicy.NewPathPolicies() with maps where the values are
pathpolicy.PathPolicy.

Without access to the pathpolicy package, the policies can be checked
using the existing ones, but no new (or modified) policies can be
created by a consumer that might have other restrictions (like,
bootc-image-builder).
